refactor: split src/utils.ts by concern
Decompose src/utils.ts (1038 LOC) into cohesive modules per #613:
runtime, conversions, format, error, async, statistics,
timer-diagnostics, timestamp, console-table, concurrency.
src/utils.ts becomes a named-re-export barrel; public API unchanged
(src/index.ts, src/bench.ts, src/task.ts untouched). No behavior change.
